Moved db require into utils.js
